+++ protected/humhub/modules/ui/widgets/DirectoryFilters.php
@@ -53,16 +53,48 @@ abstract class DirectoryFilters extends Widget
     {
         $this->initDefaultFilters();
 
-        $this->addFilter('reset', [
-            'type' => 'info',
-            'wrapperClass' => 'col-lg-2 form-search-without-info',
-            'info' => Html::a(Yii::t('UiModule.base', 'Reset filters'), [$this->pageUrl], ['class' => 'form-search-reset']),
-            'sortOrder' => 10000,
-        ]);
-
         parent::init();
 
+        // Init actions only after all filters are added by the EVENT_INIT
+        $this->initActions();
+
         ArrayHelper::multisort($this->filters, 'sortOrder');
+    }
+
+    public function initActions(): void
+    {
+        // Find min sort to put the actions after the first filter
+        $minSortOrder = null;
+        foreach ($this->filters as $data) {
+            if (isset($data['sortOrder']) && ($minSortOrder === null || $minSortOrder > $data['sortOrder'])) {
+                $minSortOrder = $data['sortOrder'];
+            }
+        }
+
+        if (count($this->filters) > 1) {
+            // Display it only to hide more filters
+            $this->addFilter('toggle-more', [
+                'type' => 'info',
+                'wrapperClass' => 'form-search-action form-search-action-toggle-more',
+                'info' => Button::light()
+                    ->icon('filter')
+                    ->action('cards.toggleMoreFilters')
+                    ->loader(false),
+                'sortOrder' => ++$minSortOrder,
+            ]);
+        }
+
+        if ($this->isFiltered()) {
+            $this->addFilter('reset', [
+                'type' => 'info',
+                'wrapperClass' => 'form-search-action form-search-action-reset',
+                'info' => Button::danger()
+                    ->icon('times')
+                    ->link([$this->pageUrl])
+                    ->tooltip(Yii::t('UiModule.base', 'Reset filters')),
+                'sortOrder' => ++$minSortOrder,
+            ]);
+        }
     }
 
     abstract protected function initDefaultFilters();
@@ -109,7 +141,7 @@ abstract class DirectoryFilters extends Widget
     public static function getDefaultFilterData(): array
     {
         return [
-            'wrapperClass' => 'col-lg-2',
+            'wrapperClass' => 'flex-fill',
             'titleClass' => 'form-search-field-info',
             'inputClass' => 'form-control',
             'beforeInput' => '',
@@ -225,6 +257,17 @@ abstract class DirectoryFilters extends Widget
         }
 
         return Yii::$app->request->get($filter, $defaultValue);
+    }
+
+    public function isFiltered(): bool
+    {
+        foreach ($this->filters as $filter => $data) {
+            if (static::getValue($filter) !== static::getDefaultValue($filter)) {
+                return true;
+            }
+        }
+
+        return false;
     }
 
 }